Add classes for representation of nodes in a graph
In order to implement a self-stabilising algorithm,
we have to subclass BaseNode class.
There are three things to do:
define variables present in every node;
define guards for every rule and
define assignments for every rule.
An example exists in the file foonode.py.
Random types defined in randomtypes.py
are useful to make initial values of the variables.