Added `run` method to the Network class
This method picks up one active node from the network
and makes a move on it.
The result is generator, which yields a dictionary
with 'prestate' and 'poststate'.
The former is state of the node before the move,
and the latter is state of the node after the move.

To count moves that brings a system to stabilisation
one can write: len(list(network.run()))